Moving stores instances to "stores"
