+++ src/hooks/useStores.js
@@ -1,5 +1,5 @@
 // import React from "react";
-// import { storesContext } from "../contexts";
+// import { storesContext } from "../stores";
 
 // hook to access stores context
 // export const useStores = () => React.useContext(storesContext);
